one_minus/hyd_yolo_training/docker_requirements.txt
ultralytics==8.1
requests==2.31
azure-keyvault-secrets==4.7
azure-identity==1.15
azureml-core==1.*
azureml-mlflow==1.*